+++ src/ui/image/facade.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 use anyhow::Result;
 use ratatui::layout::Rect;
 
-use crate::config::{Config, ImageMethod};
+use crate::config::album_art::ImageMethod;
+use crate::config::Config;
 use crate::shared::image::ImageProtocol;
 
 use super::{iterm2::Iterm2, kitty::Kitty, Backend};
@@ -37,12 +38,14 @@ impl AlbumArtFacade {
     pub fn new(config: &Config) -> Self {
         let max_size = config.album_art.max_size_px;
         let bg_color = config.theme.background_color;
+        let valign = config.album_art.vertical_align;
+        let halign = config.album_art.horizontal_align;
         let proto = match config.album_art.method.into() {
-            ImageProtocol::Kitty => ImageState::Kitty(Kitty::new(max_size, bg_color)),
+            ImageProtocol::Kitty => ImageState::Kitty(Kitty::new(max_size, bg_color, halign, valign)),
             ImageProtocol::UeberzugWayland => ImageState::Ueberzug(Ueberzug::new(Layer::Wayland, max_size)),
             ImageProtocol::UeberzugX11 => ImageState::Ueberzug(Ueberzug::new(Layer::X11, max_size)),
-            ImageProtocol::Iterm2 => ImageState::Iterm2(Iterm2::new(max_size, bg_color)),
-            ImageProtocol::Sixel => ImageState::Sixel(Sixel::new(max_size, bg_color)),
+            ImageProtocol::Iterm2 => ImageState::Iterm2(Iterm2::new(max_size, bg_color, halign, valign)),
+            ImageProtocol::Sixel => ImageState::Sixel(Sixel::new(max_size, bg_color, halign, valign)),
             ImageProtocol::None => ImageState::None,
         };
         Self {
@@ -54,19 +57,8 @@ impl AlbumArtFacade {
     }
 
     pub fn show_default(&mut self) -> Result<()> {
-        self.current_album_art = None;
-        IS_SHOWING.store(true, Ordering::Relaxed);
-
-        let data = Arc::clone(&self.default_album_art);
-        log::debug!(bytes = data.len(), area:? = self.last_size; "Displaying default image");
-
-        match &mut self.image_state {
-            ImageState::Kitty(kitty) => kitty.show(data, self.last_size),
-            ImageState::Ueberzug(ueberzug) => ueberzug.show(data, self.last_size),
-            ImageState::Iterm2(iterm2) => iterm2.show(data, self.last_size),
-            ImageState::Sixel(s) => s.show(data, self.last_size),
-            ImageState::None => Ok(()),
-        }
+        self.current_album_art = Some(Arc::clone(&self.default_album_art));
+        self.show_current()
     }
 
     pub fn show_current(&mut self) -> Result<()> {